requirements.txt
streamlit
numpy
pandas
torch
xgboost
biopython